requirements.txt
BeautifulSoup
python-dateutil